todo-django-web-app-rest-api/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
mysqlclient==2.2.4
pillow==10.2.0
platformdirs==4.2.0
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
tzdata==2024.1
